+++ evo1/seq/crystal_cavern.py
@@ -2,7 +2,7 @@ from engine.seq import SeqList
 from engine.mathlib import Facing, Vec2
 from evo1.maps import GetAStar
 from evo1.move2d import SeqGrabChest, SeqMove2D, SeqZoneTransition, SeqHoldInPlace, SeqManualUntilClose
-from evo1.atb import SeqATBmove2D
+from evo1.atb import SeqATBmove2D, SeqATBCombatManual
 from evo1.memory import MapID
 
 _cavern_astar = GetAStar(MapID.CRYSTAL_CAVERN)
@@ -18,7 +18,9 @@ class CrystalCavern(SeqList):
                 SeqATBmove2D(name="Move to trigger", coords=_cavern_astar.calculate(start=Vec2(18, 38), goal=Vec2(54, 36))),
                 SeqHoldInPlace(name="Trigger plate", target=Vec2(54, 36.5), timeout_in_s=0.5),
                 SeqATBmove2D(name="Move to boss", coords=_cavern_astar.calculate(start=Vec2(54, 36), goal=Vec2(49, 9))),
-                # TODO: Trigger fight against Kefka's ghost (interact with crystal)
+                # Trigger fight against Kefka's ghost (interact with crystal)
+                # TODO: Doesn't fully work (doesn't detect start of battle correctly)
+                #SeqATBCombatManual(name="Kefka's Ghost", wait_for_battle=True),
                 # TODO: Fight against Kefka's ghost (need to implement a smarter combat function to avoid the boss counter)
                 SeqManualUntilClose(name="Kefka's Ghost", target=Vec2(7, 10)),
                 # TODO: Grab the crystal and become 3D